liked videos section working

+++ src/pages/Room.jsx
@@ -2,17 +2,19 @@ import React, { useEffect, useState } from "react";
 import Editor from "../components/Editor";
 import { useNavigate, useParams } from "react-router-dom";
 import Canvas from "./Canvas";
-import { databases } from "../appwrite/appwriteConfig";
+import { account, databases } from "../appwrite/appwriteConfig";
 import logo from "../assets/images/logo.png";
 import { AiFillLike } from "react-icons/ai";
 import { MdDelete } from "react-icons/md";
 import { BsShareFill } from "react-icons/bs";
+import { ToastContainer, toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 const Room = () => {
   const params = useParams();
-  const navigate = useNavigate()
-  const userId = localStorage.getItem("userId");
+  const navigate = useNavigate();
   const [editor, setEditor] = useState(true);
   const [liked, setLiked] = useState(false);
+  const [userId, setUserId] = useState();
   const [project, setProject] = useState({
     name: "",
     src: {
@@ -27,7 +29,7 @@ const Room = () => {
     noOfLikes: 0,
   });
   const [code, setCode] = useState({
-    html: "hello world\n\n\n\n",
+    html: "\n\n\n\n",
     css: "*{\n   margin:0;\n   padding:0;\n}\n",
     js: `console.log("hello world");\n\n\n\n`,
     title: project?.name,
@@ -49,6 +51,7 @@ const Room = () => {
           private: response.private,
           likes: response.likes,
           noOfLikes: response.noOfLikes,
+          userId: response.userId,
         });
         setCode(JSON.parse(response?.src));
         console.log(JSON.parse(response?.src));
@@ -59,20 +62,22 @@ const Room = () => {
     );
   }
   function deleteProject() {
-    const promise = databases.deleteDocument(
-      process.env.REACT_APP_DB_ID,
-      process.env.REACT_APP_PROJECTS_COLLECTION_ID,
-      params.projectId
-    );
-    promise.then(
-      (response) => {
-        console.log(response);
-        navigate("/home")
-      },
-      (err) => {
-        console.log(err);
-      }
-    );
+    if (userId === project?.userId) {
+      const promise = databases.deleteDocument(
+        process.env.REACT_APP_DB_ID,
+        process.env.REACT_APP_PROJECTS_COLLECTION_ID,
+        params.projectId
+      );
+      promise.then(
+        (response) => {
+          console.log(response);
+          navigate("/home/" + userId);
+        },
+        (err) => {
+          console.log(err);
+        }
+      );
+    }
   }
   function saveProject(code) {
     const promise = databases.updateDocument(
@@ -120,8 +125,27 @@ const Room = () => {
       setLiked(false);
     }
   }, [project?.likes, userId]);
+  async function removeFromYourLiked() {
+    const yourLiked = await getUserInfo();
+    const removedLiked = yourLiked?.filter((c) => {
+      return c !== userId;
+    });
+    const promise = databases.updateDocument(
+      process.env.REACT_APP_DB_ID,
+      process.env.REACT_APP_USERS_COLLECTION_ID,
+      userId,
+      {
+        yourLiked: removedLiked,
+      }
+    );
+    promise.then(
+      (res) => console.log(res),
+      (err) => console.log(err)
+    );
+  }
   async function removeIdFromLiked(likesArr) {
-    const likes = likesArr.filter((c) => {
+    await removeFromYourLiked();
+    const likes = likesArr?.filter((c) => {
       return c !== userId;
     });
     const promise = databases.updateDocument(
@@ -146,7 +170,43 @@ const Room = () => {
       }
     );
   }
-  function addIdInLiked(likesArr) {
+  async function getUserInfo() {
+    let yourLiked = [];
+    const promise = databases.getDocument(
+      process.env.REACT_APP_DB_ID,
+      process.env.REACT_APP_USERS_COLLECTION_ID,
+      userId
+    );
+    promise.then(
+      (res) => {
+        console.log(res.yourLiked);
+        yourLiked = res?.yourLiked;
+      },
+      (err) => console.log(err)
+    );
+    return yourLiked;
+  }
+  async function addToYourLiked() {
+    console.log(getUserInfo());
+    const yourLiked = await getUserInfo();
+    console.log(yourLiked);
+    yourLiked?.push(params?.projectId);
+    console.log(yourLiked);
+    const promise = databases.updateDocument(
+      process.env.REACT_APP_DB_ID,
+      process.env.REACT_APP_USERS_COLLECTION_ID,
+      userId,
+      {
+        yourLiked: yourLiked,
+      }
+    );
+    promise.then(
+      (res) => console.log(res),
+      (err) => console.log(err)
+    );
+  }
+  async function addIdInLiked(likesArr) {
+    await addToYourLiked();
     likesArr.push(userId);
     console.log(likesArr);
     const promise = databases.updateDocument(
@@ -184,15 +244,38 @@ const Room = () => {
     navigator.clipboard.writeText(window.location.href).then(
       (response) => {
         console.log("Copied link to clipboard");
+        toast("Copied link to clipboard", {
+          position: "top-right",
+          autoClose: 5000,
+          hideProgressBar: false,
+          closeOnClick: true,
+          pauseOnHover: true,
+          draggable: true,
+          progress: undefined,
+          theme: "dark",
+        });
       },
       (err) => {
         console.log(err);
       }
     );
   }
+  useEffect(() => {
+    const promise = account.get();
+    promise.then(
+      (res) => {
+        setUserId(res?.$id);
+        console.log(res);
+      },
+      (err) => console.log(err)
+    );
+  }, []);
   return (
-    <div className="w-screen h-screen grid grid-cols-[230px,1fr] relative ">
-      <div className="aside sticky top-0 left-0 z-50 flex flex-col justify-between bg-slate-800 w-full h-screen text-white p-2">
+    <div className="w-screen h-screen grid sm:grid-cols-[230px,1fr] relative ">
+      <div className="fixed z-[100]">
+        <ToastContainer />
+      </div>
+      <div className="aside sticky top-0 left-0 z-50 sm:flex flex-col justify-between dark:bg-slate-950 w-full h-screen text-slate-950 dark:text-white p-2 border-r hidden ">
         <div>
           <div className="w-full h-16 flex items-center mb-4 gap-2 mt-2 border-b pb-2 justify-center">
             <img
@@ -211,8 +294,8 @@ const Room = () => {
           <div className="mb-4 flex w-full justify-end">
             <span className="flex items-center gap-1">
               <AiFillLike
-                className={`inline text-xl h-full ${
-                  liked ? "text-red-500" : "text-white"
+                className={`inline text-xl h-full cursor-pointer ${
+                  liked ? "text-red-500" : "dark:text-white "
                 } `}
                 onClick={likeClickHandler}
               />
@@ -259,17 +342,24 @@ const Room = () => {
             </li>
           </ul>
           <div className="flex items-center mt-4 mb-4 gap-1">
-            <MdDelete className="inline-flex text-3xl h-full w-auto  items-center justify-center p-1 border rounded-md cursor-pointer" onClick={deleteProject}  />
+            {console.log(userId + project?.userId)}
+            {userId === project?.userId && (
+              <MdDelete
+                className="inline-flex text-3xl h-full w-auto  items-center justify-center p-1 border rounded-md cursor-pointer"
+                onClick={deleteProject}
+              />
+            )}
             <div
               className=" p-2 bg-blue-700 rounded-lg text-center w-full cursor-pointer"
               onClick={shareClickHandler}
             >
-              <BsShareFill className="inline-flex mr-2 text-lg"/>Share
+              <BsShareFill className="inline-flex mr-2 text-lg" />
+              Share
             </div>
           </div>
         </div>
       </div>
-      <main>
+      <main className="">
         {editor ? (
           <Editor code={code} src={project?.src} setCode={setCode} />
         ) : (
